feat: remove vim-superman
Because I also use vim-root, which would change directory to the root of
the project, the path to the local man page that I wanted to preview
using vim would be incorrect. I.e., if I used `man -l ./<man page>`,
once vim-root changed the cwd, `./<man page>` could not be found.

With vim-superman, there's no easy way of correcting this. I'd have to
fork the repo and disable vim-root in the superman plugin and so on.

So, I've just reverted to the classic way of viewing man pages with vim
by piping man output to it. This is done in my bashrc.